requirements.txt
scikit-learn==1.3.0
pdfminer.six==20221105
joblib==1.3.2
pdfminer.six==20221105
scikit-learn==1.3.0
joblib==1.3.2
pandas==2.2.2
